chore: Exclude 'tests' directory from linter issues and fixed some other import issues
