+++ occupi-backend/pkg/models/request.go
@@ -1,10 +1,5 @@
 package models
 
-import (
-	"github.com/COS301-SE-2024/occupi/occupi-backend/pkg/utils"
-	"github.com/go-playground/validator/v10"
-)
-
 // expected user structure from api requests
 type RequestUser struct {
 	Email      string `json:"email" binding:"required,email"`
@@ -21,16 +16,4 @@ type RequestUserOTP struct {
 type ErrorMsg struct {
 	Field   string `json:"field"`
 	Message string `json:"message"`
-}
-
-func GetErrorMsg(fe validator.FieldError) string {
-	switch fe.Tag() {
-	case "required":
-		return "The " + utils.LowercaseFirstLetter(fe.Field()) + " field is required"
-	case "email":
-		return "The " + fe.Field() + " field must be a valid email address"
-	case "min":
-		return "The " + fe.Field() + " field must be greater than " + fe.Param()
-	}
-	return "The " + fe.Field() + " field is invalid"
 }
